Add initial RTree implementation
After looking at generated graphs, Reduced Ordered BDDs are really not a great thing for SAT with addition and other arithmetic operations.

The data structure that corresponds to what we want to do (fast comparison with many non-ordered boxes) is actually an R-Tree. This is a first implementation with really bad splitting.