Add a new learning style
The idea is to have a tree for the learned boxes where each non-leaf node has an unbounded number of children, each of which is determined by a decision and phase. This seems to lead in less costly containment test in the average case as we can cut off some branches like an R-tree but have no problem with the bounding boxes getting uselessly large.